preprocessor/requirements.txt
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.30.0
accelerate>=0.20.0
controlnet_aux>=0.0.7
safetensors>=0.4.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
Pillow>=9.0.0
numpy>=1.24.0
